update documentation for load_balancing_tool
