+++ tools/load_balancing_tool/optimize_model.py
@@ -31,27 +31,29 @@ def solver_factory(data):
         raise KeyError("totaltasks not found in data")
 
     layout = data['layout']
-    if layout in globals():
-        solverclass = globals()[layout]
-    else:
-        sp = layout.rsplit('.', 1)
+    sp = layout.rsplit('.', 1)
+    try:
         if len(sp) > 1:
-            try:
-                layout_module = importlib.import_module(sp[0])
-                layout = sp[1]
-            except ImportError, e:
-                logger.critical("ERROR: cannot import %s\n" % sp[0])
-                raise e
-        try:
-            solverclass = getattr(layout_module, layout)
-        except KeyError, e:
-            logger.critical("ERROR: layout class %s not found\n" % layout)
-            raise e
+            layout_module = importlib.import_module(sp[0])
+            layout = sp[1]
+        else:
+            import layouts
+            layout_module = layouts
+    except ImportError, e:
+        logger.critical("ERROR: cannot import %s\n", sp[0])
+        raise e
+
+    try:
+        solverclass = getattr(layout_module, layout)
+    except KeyError, e:
+        logger.critical("ERROR: layout class %s not found in %s\n",
+                        layout, layout_module)
+        raise e
 
     solver = solverclass()
 
     for c in solver.get_required_components():
-        assert(data.has_key(c), "ERROR: component %s not found in data" % c)
+        assert data.has_key(c), "ERROR: component %s not found in data" % c
 
     solver.set_data(data)
     return solver
@@ -75,13 +77,19 @@ class ModelData:
         for j in self.ntasks:
             if j % self.blocksize:
                 logger.warning("WARNING: %s pe %d not divisible by "
-                               "blocksize %d. Results may be invalid\n"
-                               % (name, j, self.blocksize))
+                               "blocksize %d. Results may be invalid\n",
+                               name, j, self.blocksize)
 
 class OptimizeModel(object):
+    STATE_UNDEFINED = 0
+    STATE_UNSOLVED = 1
+    STATE_SOLVED_OK = 2
+    STATE_SOLVED_BAD = 3
+    states = ['Undefined', 'Unsolved', 'Solved', 'No Solution']
+
     def __init__(self):
         self.models = {}
-        self.status = pulp.constants.LpStatusUndefined
+        self.state = self.STATE_UNDEFINED
 
     def set_data(self, data_dict):
         """
@@ -96,6 +104,7 @@ class OptimizeModel(object):
                  }
 
         data is extrapolated as needed for n=1 and n=totaltasks
+        sets state to STATE_UNSOLVED
         """
         # get deep copy, because we need to divide ntasks by blocksize
         self.maxtasks = data_dict['totaltasks']
@@ -132,7 +141,7 @@ class OptimizeModel(object):
                 m.extrapolated.append(True)
 
         self.check_requirements()
-        self.status = pulp.constants.LpStatusUndefined
+        self.state = self.STATE_UNSOLVED
 
     def add_model_constraints(self):
         """
@@ -140,6 +149,8 @@ class OptimizeModel(object):
         This should be the same for any layout so is provided in base class
         Assumes cost variables are 'Txxx' and ntask variables are 'Nxxx'
         """
+        assert self.state != self.STATE_UNDEFINED,\
+               "set_data() must be called before add_model_constraints()"
         for k in self.get_required_components():
             m = self.models[k]
             tk = 'T' + k.lower() # cost(time) key
@@ -152,19 +163,20 @@ class OptimizeModel(object):
                                          (k.lower(), slope, k.lower(),
                                           m.cost[i] - slope * m.ntasks[i])])
                 if slope > 0:
-                    logger.warning("WARNING: Nonconvex cost function for "
-                                   "model %s. Please")
-                    logger.warning("Review costs to ensure data is correct "
-                                   "(--graph_models or --print_models)")
+                    logger.warning("WARNING: Nonconvex cost function for model "
+                                   "%s. Review costs to ensure data is correct "
+                                   "(--graph_models or --print_models)", k)
+
                     break
                 if slope == 0:
                     break
 
-        self.status = pulp.constants.LpStatusNotSolved
-
     def get_required_components(self):
         """
-        Should be overridden by derived class
+        Should be overridden by derived class. Return a list of required
+        components (capitalized) used in the layout.
+
+        Example: return ['ATM', 'LND', 'ICE']
         """
         return []
 
@@ -175,7 +187,7 @@ class OptimizeModel(object):
         """
         for r in self.get_required_components():
             if r not in self.models:
-                logger.critical("Data for component %s not available" % r)
+                logger.critical("Data for component %s not available", r)
 
     def write_timings(self, fd=sys.stdout, level=logging.DEBUG):
         """
@@ -183,6 +195,8 @@ class OptimizeModel(object):
         Can be used to check that the data provided to the
         model is reasonable. Also see graph_costs()
         """
+        assert self.state != self.STATE_UNDEFINED,\
+               "set_data() must be called before write_timings()"
         for k in self.models:
             m = self.models[k]
             message = "***%s***" % k
@@ -208,6 +222,8 @@ class OptimizeModel(object):
 
         If matplotlib is not available, nothing will happen
         """
+        assert self.state != self.STATE_UNDEFINED,\
+               "set_data() must be called before graph_costs()"
         try:
             import matplotlib.pyplot as pyplot
         except ImportError, e:
@@ -252,13 +268,14 @@ class OptimizeModel(object):
     def optimize(self):
         """
         Run the optimization.
-        Must set self.status using LpStatus object
-           1  "Optimal"       LpStatusOptimal
-           0  "Not Solved"    LpStatusNotSolved
-          -1  "Infeasible"    LpStatusInfeasible
-          -2  "Unbounded"     LpStatusUnbounded
-          -3  "Undefined"     LpStatusUndefined
-        Returns LpStatus
+        Must set self.state using LpStatus object:
+              LpStatusOptimal    -> STATE_SOLVED_OK
+              LpStatusNotSolved  -> STATE_UNSOLVED
+              LpStatusInfeasible -> STATE_SOLVED_BAD
+              LpStatusUnbounded  -> STATE_SOLVED_BAD
+              LpStatusUndefined  -> STATE_UNDEFINED
+              -- use self.set_state(lpstatus) --
+        Returns state
 
         If solved, then solution will be stored in self.X dictionary, indexed
         by variable name. Suggested convention:
@@ -273,12 +290,32 @@ class OptimizeModel(object):
 
     def get_solution(self):
         """
-        Return a dictionary of the solution variables, can be overridden
+        Return a dictionary of the solution variables, can be overridden.
+        Default implementation returns values in self.X
         """
+        assert self.state == self.STATE_SOLVED_OK,\
+               "solver failed, no solution available"
         retval = {}
-        for k in self.X:
-            retval[k] = self.X[k].varValue
+        if hasattr(self,'X') and isinstance(self.X, dict):
+            for k in self.X:
+                retval[k] = self.X[k].varValue
         return retval
+
+    def set_state(self, lpstatus):
+        if lpstatus == pulp.constants.LpStatusOptimal:
+            self.state = self.STATE_SOLVED_OK
+        elif lpstatus == pulp.constants.LpStatusNotSolved:
+            self.state = self.STATE_UNSOLVED
+        elif lpstatus == pulp.constants.LpStatusUndefined:
+            self.state = self.STATE_UNDEFINED
+        else:
+            self.state = self.STATE_SOLVED_BAD
+
+    def get_state(self):
+        return self.state
+
+    def get_state_string(self, state):
+        return self.states[state]
 
     def get_all_variables(self):
         return self.prob.X
@@ -297,7 +334,7 @@ class OptimizeModel(object):
         from distutils.spawn import find_executable
         from xml.etree import ElementTree as ET
         from CIME.utils import run_cmd
-        logger.info("Writing pe node info to %s" % pefilename)
+        logger.info("Writing pe node info to %s", pefilename)
         root = ET.Element('config_pes')
         grid = ET.SubElement(root, 'grid')
         grid.set('name', 'any')
@@ -319,159 +356,6 @@ class OptimizeModel(object):
             node = ET.SubElement(rootpe_node, 'rootpe_' + k)
             node.text = str(roots[k])
         xmllint = find_executable("xmllint")
-        import pdb; pdb.set_trace()
         if xmllint is not None:
             run_cmd("%s --format --output %s -" % (xmllint, pefilename),
                     input_str=ET.tostring(root))
-
-class IceLndAtmOcn(OptimizeModel):
-    """
-    Optimized the problem based on the Layout
-              ____________________
-             | ICE  |  LND  |     |
-             |______|_______|     |
-             |              | OCN |
-             |    ATM       |     |
-             |______________|_____|
-
-      Min T
-      s.t.  T[ice]      <= T1
-            T[lnd]      <= T1
-            T1 + T[atm] <= T
-            T[ocn]      <= T
-
-            NB[c]        >= 1 for c in [ice,lnd,ocn,atm]
-            NB[ice] + NB[lnd] <= NB[atm]
-            atm_blocksize*NB[atm] + ocn_blocksize*NB[ocn] <= TotalTasks
-            (NB[*] is number of processor blocks)
-
-            T[c]        >= C[c]_{i} - NB[c]_{i} *
-                       (C[c]_{i+1} - C[c]_{i}) / (NB[c]_{i+1} - NB[c]_{i})
-                       + NB[c] * (C[c]_{i+1} - C[c]_{i})
-                                               / (NB[c]_{i+1} - NB[c]_{i}),
-                        i=1..ord(NB), c in [ice,lnd,ocn,atm]
-
-    These assumptions are checked when solver is initialized
-      . Assuming cost is monotonic decreasing vs ntasks
-      . Assuming perfect scalability for ntasks < tasks[0]
-      . Assuming same scalability factor for ntasks > ntasks[last] as for
-                              last two data points
-
-    Returns LpStatus
-    If solved, then solution will be stored in self.X dictionary, indexed
-    by variable name. Suggested convention:
-        'Tice', 'Tlnd', ... for cost per component
-        'Nice', 'Nlnd', ... for ntasks per component
-        'NBice', 'NBlnd', ... for number of blocks per component
-    """
-    def get_required_components(self):
-        return ['LND', 'ICE', 'ATM', 'OCN']
-
-    def optimize(self):
-        """
-        Run the optimization.
-        Must set self.status using LpStatus object
-           1  "Optimal"       LpStatusOptimal
-           0  "Not Solved"    LpStatusNotSolved
-          -1  "Infeasible"    LpStatusInfeasible
-          -2  "Unbounded"     LpStatusUnbounded
-          -3  "Undefined"     LpStatusUndefined
-        """
-        assert(self.status == pulp.constants.LpStatusUndefined,
-               "set_data() must be called before optimize()!")
-        self.atm = self.models['ATM']
-        self.lnd = self.models['LND']
-        self.ice = self.models['ICE']
-        self.ocn = self.models['OCN']
-        self.real_variables = ['TotalTime', 'T1', 'Tice', 'Tlnd', 'Tatm',
-                               'Tocn']
-        self.integer_variables = ['NBice', 'NBlnd', 'NBatm', 'NBocn',
-                                  'Nice', 'Nlnd', 'Natm', 'Nocn']
-        self.X = {}
-        X = self.X
-        self.prob = pulp.LpProblem("Minimize ACME time cost", pulp.LpMinimize)
-        for rv in self.real_variables:
-            X[rv] = pulp.LpVariable(rv, lowBound=0)
-
-        for iv in self.integer_variables:
-            X[iv] = pulp.LpVariable(iv, lowBound=1, cat=pulp.LpInteger)
-
-
-        # cost function
-        self.prob += X['TotalTime']
-
-        #constraints
-        self.constraints = []
-        # Layout-dependent constraints. Choosing another layout to model
-        # will require editing these constraints
-        self.constraints.append([X['Tice'] - X['T1'] <= 0, "Tice - T1 == 0"])
-        self.constraints.append([X['Tlnd'] - X['T1'] <= 0, "Tlnd - T1 == 0"])
-        self.constraints.append([X['T1'] + X['Tatm'] - X['TotalTime'] <= 0,
-                                 "T1 + Tatm - TotalTime <= 0"])
-        self.constraints.append([X['Tocn'] - X['TotalTime'] <= 0,
-                                 "Tocn - TotalTime == 0"])
-        self.constraints.append([X['Nice'] + X['Nlnd'] - X['Natm'] == 0,
-                                 "Nice + Nlnd - Natm == 0"])
-        self.constraints.append([X['Natm'] + X['Nocn'] == self.maxtasks,
-                                 "Natm + Nocn <= %d" % (self.maxtasks)])
-        self.constraints.append([self.atm.blocksize * X['NBatm'] - X['Natm'] == 0,
-                                 "Natm = %d * NBatm" % self.atm.blocksize])
-        self.constraints.append([self.ice.blocksize * X['NBice'] - X['Nice'] == 0,
-                                 "Nice = %d * NBice" % self.ice.blocksize])
-        self.constraints.append([self.lnd.blocksize * X['NBlnd'] - X['Nlnd'] == 0,
-                                 "Nlnd = %d * NBlnd" % self.lnd.blocksize])
-        self.constraints.append([self.ocn.blocksize * X['NBocn'] - X['Nocn'] == 0,
-                                 "Nocn = %d * NBocn" % self.ocn.blocksize])
-
-        # These are the constraints based on the timing data.
-        # They should be the same no matter what the layout of the components.
-        self.add_model_constraints()
-
-        for c, s in self.constraints:
-            self.prob += c, s
-
-        # Write the program to file and solve (using coin-cbc)
-        self.prob.writeLP("IceLndAtmOcn_model.lp")
-        self.prob.solve()
-        self.status = self.prob.status
-        return pulp.LpStatus[self.prob.status]
-
-    def get_solution(self):
-        """
-        Return a dictionary of the solution variables.
-        """
-        return {'NBLOCKS_ICE':self.X['NBice'].varValue,
-                'NBLOCKS_LND':self.X['NBlnd'].varValue,
-                'NBLOCKS_ATM':self.X['NBatm'].varValue,
-                'NBLOCKS_OCN':self.X['NBocn'].varValue,
-                'NTASKS_ICE':self.X['Nice'].varValue,
-                'NTASKS_LND':self.X['Nlnd'].varValue,
-                'NTASKS_ATM':self.X['Natm'].varValue,
-                'NTASKS_OCN':self.X['Nocn'].varValue,
-                'NTASKS_TOTAL':self.maxtasks,
-                'COST_ICE':self.X['Tice'].varValue,
-                'COST_LND':self.X['Tlnd'].varValue,
-                'COST_ATM':self.X['Tatm'].varValue,
-                'COST_OCN':self.X['Tocn'].varValue,
-                'COST_TOTAL':self.X['TotalTime'].varValue}
-
-    def write_pe_file(self, pefilename):
-        """
-        Write out a pe_file that can be used to implement the
-        optimized layout
-        """
-        natm = int(self.X['Natm'].varValue)
-        nlnd = int(self.X['Nlnd'].varValue)
-        nice = int(self.X['Nice'].varValue)
-        nocn = int(self.X['Nocn'].varValue)
-        ntasks = {'atm':natm, 'lnd':nlnd, 'rof':1, 'ice':nice,
-                  'ocn':nocn, 'glc':1, 'wav':1, 'cpl':1}
-        roots = {'atm':0, 'lnd':nice, 'rof':0, 'ice':0,
-                 'ocn':natm, 'glc':0, 'wav':0, 'cpl':0}
-        nthrds = {}
-        for c in ['atm', 'lnd', 'rof', 'ice', 'ocn', 'glc', 'wav', 'cpl']:
-            if c.upper() in self.models:
-                nthrds[c] = self.models[c.upper()].nthrds
-            else:
-                nthrds[c] = 1
-        self.write_pe_template(pefilename, ntasks, nthrds, roots)
