+++ tools/load_balancing_tool/tests/test.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0"?>
+<config_pes>
+  <grid name="any">
+    <mach name="any">
+      <pes compset="any" pesize="">
+        <ntasks>
+          <ntasks_glc>1</ntasks_glc>
+          <ntasks_cpl>1</ntasks_cpl>
+          <ntasks_ice>872</ntasks_ice>
+          <ntasks_ocn>32</ntasks_ocn>
+          <ntasks_lnd>120</ntasks_lnd>
+          <ntasks_atm>992</ntasks_atm>
+          <ntasks_wav>1</ntasks_wav>
+          <ntasks_rof>1</ntasks_rof>
+        </ntasks>
+        <nthrds>
+          <nthrds_glc>1</nthrds_glc>
+          <nthrds_cpl>1</nthrds_cpl>
+          <nthrds_ocn>1</nthrds_ocn>
+          <nthrds_atm>1</nthrds_atm>
+          <nthrds_ice>1</nthrds_ice>
+          <nthrds_rof>1</nthrds_rof>
+          <nthrds_wav>1</nthrds_wav>
+          <nthrds_lnd>1</nthrds_lnd>
+        </nthrds>
+        <rootpe>
+          <rootpe_glc>0</rootpe_glc>
+          <rootpe_cpl>0</rootpe_cpl>
+          <rootpe_ice>0</rootpe_ice>
+          <rootpe_ocn>992</rootpe_ocn>
+          <rootpe_lnd>872</rootpe_lnd>
+          <rootpe_atm>0</rootpe_atm>
+          <rootpe_wav>0</rootpe_wav>
+          <rootpe_rof>0</rootpe_rof>
+        </rootpe>
+      </pes>
+    </mach>
+  </grid>
+</config_pes>
